refactor: extract getPrFromItem helper, split usePaneMode and main-input
- Add getPrFromItem() to types.ts, replacing 6 identical redundant ternaries
- Split usePaneMode into 4 focused hooks: usePaneMode (slim), useDiffState,
  useCommentState, useReviewConfirmState
- Replace queueMicrotask with React's idiomatic "store previous value" pattern
- Split 1090-line main-input.ts into 7 focused modules with barrel re-export:
  input-types, branch-picker-input, confirm-delete-input, diff-file-list-input,
  diff-viewer-input, confirm-input, sidebar-input

Pure refactor — no behavioral changes.
