+++ apps/cli/src/types.ts
@@ -2,7 +2,12 @@ import type { PullRequestInfo } from '@kirby/vcs-core';
 
 export type Focus = 'sidebar' | 'terminal';
 
-export type PaneMode = 'terminal' | 'pr-detail' | 'diff' | 'diff-file' | 'confirm';
+export type PaneMode =
+  | 'terminal'
+  | 'pr-detail'
+  | 'diff'
+  | 'diff-file'
+  | 'confirm';
 
 export type ReviewCategory = 'needs-review' | 'waiting' | 'approved';
 
@@ -16,7 +21,17 @@ export type SidebarItem =
       conflictCount?: number;
     }
   | { kind: 'orphan-pr'; pr: PullRequestInfo; running?: boolean }
-  | { kind: 'review-pr'; pr: PullRequestInfo; category: ReviewCategory; running?: boolean };
+  | {
+      kind: 'review-pr';
+      pr: PullRequestInfo;
+      category: ReviewCategory;
+      running?: boolean;
+    };
+
+/** Extract the PR from any sidebar item kind. */
+export function getPrFromItem(item: SidebarItem): PullRequestInfo | undefined {
+  return item.pr;
+}
 
 export interface AgentSession {
   name: string;
